Finish algorithm implementation
This is able to generate a compelling citation graph in a reasonable
amount of time.  The code quality isn't great, I'm calling .clone()
*way* too much, but that'll be fixed in good time.

The next step is to create a CLI so not everything is hard-coded.

It'd also be good to write tests to make sure that this continues to
work in the future and to encourage me to encapsulate some
functionality.