Implement IDs from BibTeX file
Since this project is about growing bibliographies in useful ways, it
made sense to me to get the IDs, here DOIs and Semantic Scholar URLs,
from a bibliography.

Takeaway for future work is that performance is not great.  If the
futures for the API calls could be dispatched all at once, I could see
that leading to big gains.