Linearize the connectivity
Previously, a change of 0.01 in connectivity would dramatically alter
the behavior of the program.  I want that to be more like 0.1, so I made
the minimum citation count be more like connectivity ** depth.

In the course of this, I found that Semantic Scholar was giving a lot of
429 Too Many Requests and 504 Gateway Timeout errors.  I slowed down the
rate limiter by 10% and set to do retrys on 504 to adjust for this.

I should add some error handling to the client, though.